pathexists() - takes a path, checks it and return 1 : 0 if it exists.
reextend() - takes argument member and extends the destination_pwd and trace_file_loc
to reflect and adjust to filees in the trash sharing the same name. 1 -> 1_re -> 1_re_re ..

deletefile - reimplemented to act when a file in the trash directory shares the same filename
to avoid overwrites. Now depenedent on reextend in these instances.